+++ src/gleam_javascript_ffi.mjs
@@ -85,7 +85,7 @@ export function newPromise(executor) {
   return new Promise((resolve) =>
     executor((value) => {
       resolve(PromiseLayer.wrap(value));
-    })
+    }),
   );
 }
 
@@ -99,36 +99,12 @@ export function then_await(promise, fn) {
 
 export function map_promise(promise, fn) {
   return promise.then((value) =>
-    PromiseLayer.wrap(fn(PromiseLayer.unwrap(value)))
+    PromiseLayer.wrap(fn(PromiseLayer.unwrap(value))),
   );
 }
 
 export function rescue(promise, fn) {
   return promise.catch((error) => fn(error));
-}
-
-class Reference {
-  constructor(value) {
-    this.value = value;
-  }
-}
-
-export function dereference(reference) {
-  return reference.value;
-}
-
-export function make_reference(value) {
-  return new Reference(value);
-}
-
-export function set_reference(ref, value) {
-  let previous = ref.value;
-  ref.value = value;
-  return previous;
-}
-
-export function reference_equal(a, b) {
-  return a === b
 }
 
 export function all_promises(...promises) {